Add Main component
